requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.7.4.post0=py39hb82d6ee_0
async-timeout=3.0.1=py_1000
attrs=21.2.0=pyhd8ed1ab_0
blinker=1.4=py39haa95532_0
brotlipy=0.7.0=py39hb82d6ee_1001
ca-certificates=2021.5.30=h5b45459_0
cachetools=4.2.2=pyhd8ed1ab_0
certifi=2021.5.30=py39hcbf5309_0
cffi=1.14.6=py39h0878f49_0
chardet=4.0.0=py39hcbf5309_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=8.0.1=pyhd3eb1b0_0
cryptography=3.4.7=py39hd8d06c1_0
google-api-core=1.31.0=pyhd8ed1ab_0
google-api-python-client=2.13.0=pyhd8ed1ab_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-httplib2=0.1.0=pyhd8ed1ab_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
googleapis-common-protos=1.53.0=py39h35db3c3_0
httplib2=0.19.1=pyhd8ed1ab_0
idna=3.1=pyhd3deb0d_0
importlib-metadata=3.10.0=py39haa95532_0
libprotobuf=3.17.2=h7755175_1
multidict=5.1.0=py39hb82d6ee_1
oauthlib=3.1.1=pyhd3eb1b0_0
openssl=1.1.1k=h8ffe710_0
packaging=21.0=pyhd8ed1ab_0
pip=21.1.3=py39haa95532_0
protobuf=3.17.2=py39h415ef7b_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pyjwt=2.1.0=py39haa95532_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py39hcbf5309_3
python=3.9.5=h6244533_3
python_abi=3.9=2_cp39
pytz=2021.1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
requests=2.26.0=pyhd8ed1ab_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyh44b312d_0
selenium=3.141.0=py39hb82d6ee_1002
setuptools=52.0.0=py39haa95532_0
simplejson=3.17.3=py39hb82d6ee_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.36.0=h2bbff1b_0
typing-extensions=3.10.0.0=hd8ed1ab_0
typing_extensions=3.10.0.0=pyha770c72_0
tzdata=2021a=h52ac0ba_0
uritemplate=3.0.1=py_0
urllib3=1.26.6=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39hcbf5309_2
wincertstore=0.2=py39h2bbff1b_0
yarl=1.6.3=py39hb82d6ee_2
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_1010
